#3647 table matches ruleset type, fixed pages
